+++ packages/commons/enums/favTagType.ts
@@ -1,0 +1,4 @@
+export enum FavTagType {
+  tag = 'tag',
+  groupTag = 'groupTag',
+}