+++ packages/commons/models/favTag.ts
@@ -1,0 +1,11 @@
+import { FavTagType } from '../enums'
+import { IPromotedTag } from './promotedTag'
+
+export interface IFavTag {
+  id: string
+  tag?: string
+  groupTag: IPromotedTag
+  type: FavTagType
+  createdAt: string
+  updatedAt: string
+}