+++ packages/state-post/src/posts.store.ts
@@ -1,5 +1,8 @@
 import { EntityState, EntityStore } from '@datorama/akita'
-import { IPost } from '@gtms/commons/models'
+import { IPost, IPostImage } from '@gtms/commons/models'
+import { parsePostOwnersAvatar } from './helpers'
+import { parseFiles } from '@gtms/commons/helpers'
+import { FileStatus } from '@gtms/commons/enums'
 import { Sorting } from '@gtms/api-post'
 
 export interface IPostsState extends EntityState<IPost, string> {
@@ -27,6 +30,36 @@ export class PostsStore extends EntityStore<IPostsState> {
         resettable: true,
       }
     )
+  }
+
+  private parseFiles(post: IPost): IPost {
+    post = parsePostOwnersAvatar(post)
+
+    if (Array.isArray(post.images)) {
+      post.images = post.images
+        .map((img) => {
+          if (img.status !== FileStatus.ready) {
+            return null
+          }
+
+          if (Array.isArray(img.files)) {
+            img.files = parseFiles(img.files)
+          }
+
+          return img
+        })
+        .filter((img) => img) as IPostImage[]
+    }
+
+    return post
+  }
+
+  akitaPreAddEntity = (post: IPost) => {
+    return this.parseFiles(post)
+  }
+
+  akitaPreUpdateEntity = (post: IPost) => {
+    return this.parseFiles(post)
   }
 }