+++ packages/api-tags/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.11.0](https://github.com/gtms-org/gtms-frontend/compare/@gtms/api-tags@0.10.0...@gtms/api-tags@0.11.0) (2020-12-27)
+
+
+### Features
+
+* GEOT-734 add to favs ([#957](https://github.com/gtms-org/gtms-frontend/issues/957)) ([771cb8d](https://github.com/gtms-org/gtms-frontend/commit/771cb8d2bae65c06e4121b3ec4db8d0ff504356e))
+
+
+
+
+
 # [0.10.0](https://github.com/gtms-org/gtms-frontend/compare/@gtms/api-tags@0.9.1...@gtms/api-tags@0.10.0) (2020-12-15)