+++ packages/state-group/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.20.3](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-group@0.20.2...@gtms/state-group@0.20.3) (2020-12-28)
+
+
+### Bug Fixes
+
+* GEOT-743: migrate to store hooks ([#958](https://github.com/gtms-org/gtms-frontend/issues/958)) ([9fac1d5](https://github.com/gtms-org/gtms-frontend/commit/9fac1d55746457d9cb22d70642891e8d230a4b35))
+
+
+
+
+
+## [0.20.2](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-group@0.20.1...@gtms/state-group@0.20.2) (2020-12-27)
+
+**Note:** Version bump only for package @gtms/state-group
+
+
+
+
+
 ## [0.20.1](https://github.com/gtms-org/gtms-frontend/compare/@gtms/state-group@0.20.0...@gtms/state-group@0.20.1) (2020-12-25)
 
 **Note:** Version bump only for package @gtms/state-group